Small change to Compendium.Gamedata to allow for  Prefabs to be serialized
