+++ Assets/Scripts/System/Utilities/Compendium/Prefabs.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d43a16a3ed8a79a45bd6db9eba0ef036
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
